apps/backend/requirements.txt
# Core API framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2

# HTTP requests and file uploads
requests==2.31.0
python-multipart==0.0.7
aiohttp==3.9.1

# Database and caching - Using compatible versions
supabase==2.4.4
supafunc==0.3.2
gotrue==1.3.1
postgrest==0.13.2

# Configuration and scheduling
python-dotenv==1.0.0
apscheduler==3.10.4

# Authentication and security
PyJWT==2.8.0
cryptography==41.0.7

# System monitoring
psutil==5.9.6

# Enhanced logging and monitoring
python-json-logger==2.0.7